No longer require the user to specify a 'friendly_name' for progressive render blocks.
Introduce the FragmentNameIterator to deterministically generate names on a page.
Assuming a single request is handled by a single thread, this should be a safe approach.
It does not support nested progreesive_render sections. I'm not clear that that's a real scenario.

The friendly_name param is now deprecated and will be ignored. This is shown to the user via
a warning when a progressive_render call specifies a friendly_name. This will be removed in v1

Fixes #10