Modify the DOM instead of setting the 'data' prop via jQuery to mark a section as 'progressive-loaded'
Defect:
This was exposed as an issue by another bug which was loading the entire page on partial load and re-loading the javascript.
The result was an infinite loop of reloading the content.

Fix:
Modifying the DOM of the element we don't replace (the main placeholder) ensure we don't run into this class of issues.